+++ releases.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
   - Fix handling of inspect on Ruby 4+.
+  - `Client#run(**options)` are passed through to `Async(**options)`.
 
 ## v0.3.1